toolkit/components/passwordmgr/test/mochitest/test_autofill_tab_between_fields.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autocomplete behavior when tabbing between form fields</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
let nsLoginInfo = SpecialPowers.wrap(SpecialPowers.Components).Constructor("@mozilla.org/login-manager/loginInfo;1",
                                                                           SpecialPowers.Ci.nsILoginInfo,
                                                                           "init");
let readyPromise = registerRunTests(1);
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">
  <form id="form1" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

async function prepareLogins(logins = []) {
  await LoginManager.removeAllLogins();

  for (let login of logins) {
    let storageAddPromise = promiseStorageChanged(["addLogin"]);
    await LoginManager.addLogin(login);
    await storageAddPromise;
  }
  let count = (await LoginManager.getAllLogins()).length;
  is(count, logins.length, "All logins were added");
}

let origin = window.location.origin;
const availableLogins = {
  "exampleUser1": new nsLoginInfo(origin, "https://autofill", null,
                              "user1", "pass1", "uname", "pword"),
}

const tests = [
  {
    name: "autofill_disabled_exact_username",
    autofillEnabled: false,
    logins: ["exampleUser1"],
    expectedAutofillUsername: "",
    expectedAutofillPassword: "",
    typeUsername: "user1",
    expectedTabbedUsername: "user1",
    expectedTabbedPassword: "",
  },
  {
    name: "autofill_enabled_exact_username",
    autofillEnabled: true,
    logins: ["exampleUser1"],
    expectedAutofillUsername: "user1",
    expectedAutofillPassword: "pass1",
    typeUsername: "user1",
    expectedTabbedUsername: "user1",
    expectedTabbedPassword: "pass1",
  },
];

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;
});

async function testResultOfTabInteractions(testData) {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.autofillForms", testData.autofillEnabled],
  ]});

  await SimpleTest.promiseFocus(window);
  let logins = testData.logins.map(name => availableLogins[name]);
  await prepareLogins(logins);

  info("recreating form");
  let processed = promiseFormsProcessed();
  recreateTree(document.getElementById("form1"));
  info("waiting for form processed");
  await processed;
  // check autofill results
  checkForm(1, testData.expectedAutofillUsername, testData.expectedAutofillPassword);
  let pword = $_(1, "pword");
  let uname = $_(1, "uname");

  SpecialPowers.wrap(pword).setUserInput("");
  SpecialPowers.wrap(uname).setUserInput("");

  info("Placing focus in the username field");
  uname.focus();

  if (testData.typeUsername) {
    info("Filling username field");
    await sendString(testData.typeUsername);
  }

  await synthesizeKey("KEY_Tab"); // blur un, focus pw
  await new Promise(resolve => SimpleTest.executeSoon(resolve));

  ok($_(1, "pword").matches("input:focus"), "pword field is focused");
  checkForm(1, testData.expectedTabbedUsername, testData.expectedTabbedPassword);

  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessed();
  await synthesizeKey("KEY_Escape");
}

for (let testData of tests) {
  let tmp = {
    async [testData.name]() {
      await testResultOfTabInteractions(testData);
    },
  };
  add_task(tmp[testData.name]);
}

</script>
</pre>
</body>
</html>
